+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Allow selecting and copying text from tool execution results (https://github.com/nbonamy/witsy/issues/421)
 - Support Perplexity Search API (https://github.com/nbonamy/witsy/issues/427)
 - Web 'Apps' (https://github.com/nbonamy/witsy/issues/431)
+- OpenAI: Hide models with explicit release dates
 - Table rendering as artifact, download as CSV and XSLX
 - HTTP Server to control Witsy (see README.md)
 - Agent trigger via webhooks (see README.md)